refactor(ai): split ai_base.py into ai/base.py + ai/gemini_client.py (Related to #1953)
Extract provider-agnostic utilities (retry constants, JSON repair) into
ai/base.py (zero google.genai imports) and Gemini/Vertex AI client into
ai/gemini_client.py. ai_base.py becomes a thin backward-compat facade that
re-exports the full public surface including capture_usage/last_usage/TUTOR_PERSONA.
39 TDD characterization tests added (all GREEN).
